feat: split out tests
